restructure db connection and folders
