+++ src/main/java/com/example/demo/daos/Dao.java
@@ -11,8 +11,8 @@ public interface Dao<T> {
     List<T> getAll() throws SQLException;
 
 //    void save(T t);
-//
-//    void insert(T t);
+
+    void create(T t) throws SQLException;
 //
 //    void update(T t);
 //
