+++ src/main/java/com/example/demo/servlets/auth/Signup.java
@@ -1,9 +1,10 @@
-package com.example.demo.servlet;
+package com.example.demo.servlets.auth;
 
 import com.example.demo.Validation;
 import com.example.demo.beans.Alert;
 import com.example.demo.beans.forms.SignupForm;
 import com.example.demo.daos.UserDao;
+import com.example.demo.servlets.DatabaseHttpServlet;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
@@ -11,11 +12,10 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
 import java.io.IOException;
+import java.sql.SQLException;
 
 @WebServlet(name = "Signup", value = "/signup")
-public class Signup extends HttpServlet {
-    public void init() {}
-
+public class Signup extends DatabaseHttpServlet {
     public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException {
         request.getRequestDispatcher("/template/auth/signup.jsp").forward(request, response);
     }
@@ -36,16 +36,24 @@ public class Signup extends HttpServlet {
                 user.setPassword(form.getPassword());
                 user.setEmail(form.getEmail());
 
-                if(user.isUnique() != 0){
-                    user.createNewUser();
+                try {
+                    if(user.isUnique() != 0){
+                        user.createNewUser();
 
-                    response.sendRedirect("index.jsp");
-                    return;
-                }
-                else {
+                        response.sendRedirect("index.jsp");
+                        return;
+                    }
+                    else {
+                        request.setAttribute(
+                                "alert",
+                                new Alert("danger", "Duplicate username/password")
+                        );
+                    }
+                } catch (SQLException e) {
+                    e.printStackTrace();
                     request.setAttribute(
                             "alert",
-                            new Alert("danger", "Duplicate username/password")
+                            new Alert("danger", "An error occurred while creating your account. Please try again later.")
                     );
                 }
             }
@@ -57,9 +65,5 @@ public class Signup extends HttpServlet {
         request.setAttribute("form", form);
 
         request.getRequestDispatcher("/template/auth/signup.jsp").forward(request, response);
-    }
-
-
-    public void destroy() {
     }
 }
